% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arch.test.R
\name{arch.test}
\alias{arch.test}
\title{Arch Test (AT) Stationarity Test}
\usage{
arch.test(data, alpha)
}
\arguments{
\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}
}
\value{
0: Non-Stationary, 1: Stationary, NA: Unable to test
}
\description{
Arch Test stationarity test.
}

