% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmented.dickey.fuller.test.R
\name{augmented.dickey.fuller.test}
\alias{augmented.dickey.fuller.test}
\title{Augmented Dickey-Fuller (ADF) Stationarity Test}
\usage{
augmented.dickey.fuller.test(data, alpha)
}
\arguments{
\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}
}
\value{
0: Non-Stationary, 1: Stationary, NA: Unable to test
}
\description{
Augmented Dickey-Fuller stationarity test.
}
