% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.test.R
\name{bootstrap.test}
\alias{bootstrap.test}
\title{Bootstrap Wavelet Packet Tests}
\usage{
bootstrap.test(data, alpha)
}
\arguments{
\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}
}
\value{
0: Non-Stationary, 1: Stationary, NA: Unable to test
}
\description{
Bootstrap wavelet packet tests.
}
