% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/bootwptos.test.R
=======
% Please edit documentation in R/bootstrap.test.R
>>>>>>> 1980ca786f9bc64960213752d59f98e3bafdc768
\name{bootwptos.test}
\alias{bootwptos.test}
\title{Bootstrap Wavelet Packet Tests}
\usage{
bootwptos.test(data, alpha)
}
\arguments{
\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}
}
\value{
0: Non-Stationary, 1: Stationary, NA: Unable to test
}
\description{
Bootstrap wavelet packet tests.
}
<<<<<<< HEAD
=======

>>>>>>> 1980ca786f9bc64960213752d59f98e3bafdc768
