% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breusch.godfrey.test.R
\name{breusch.godfrey.test}
\alias{breusch.godfrey.test}
\title{Breusch-Godfrey (BG) Stationarity Test}
\usage{
breusch.godfrey.test(data, alpha)
}
\arguments{
\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}
}
\value{
0: Non-Stationary, 1: Stationary, NA: Unable to test
}
\description{
Breusch-Godfrey stationarity test.
}
