% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breusch.pagan.test.R
\name{breusch.pagan.test}
\alias{breusch.pagan.test}
\title{Breusch-Pagan (BP) Stationarity Test}
\usage{
breusch.pagan.test(data, alpha)
}
\arguments{
\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}
}
\value{
0: Non-Stationary, 1: Stationary, NA: Unable to test
}
\description{
Breusch-Pagan stationarity test.
}

