% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdw.test.R
\name{cdw.test}
\alias{cdw.test}
\title{Chli and De Wilde stability algorithm}
\usage{
cdw.test(data, alpha)
}
\arguments{
\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}
}
\value{
0: Non-Stationary, 1: Stationary
}
\description{
Procedure to test for time series stability described in
Chli, M. and de Wilde, P. (2009). Convergence and knowledge processing in
multi-agent systems. Springer.
}
