% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpm.test.R
\name{cpm.test}
\alias{cpm.test}
\title{Change Point Stationarity Test}
\usage{
cpm.test(data, type)
}
\arguments{
\item{data}{Time series data}

\item{type}{"Student", "Bartlett", "GLR", "Exponential", "GLRAdjusted",
"FET", "Mann-Whitney", "Mood", "Lepage", "Kolmogorov-Smirnov",
"Cramer-von-Mises"}
}
\value{
0: Non-Stationary, 1: Stationary, NA: Unable to test
}
\description{
Change Point changes stationarity test.
}
