% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dickey.fuller.test.R
\name{dickey.fuller.test}
\alias{dickey.fuller.test}
\title{Dickey-Fuller (DF) Stationarity Test}
\usage{
dickey.fuller.test(data, alpha)
}
\arguments{
\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}
}
\value{
0: Non-Stationary, 1: Stationary, NA: Unable to test
}
\description{
Dickey-Fuller stationarity test.
}
