% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2steps.test.R
\name{dt2steps.test}
\alias{dt2steps.test}
\title{Decision Tree 2-Steps Stationarity Test}
\usage{
dt2steps.test(data, alpha, mode)
}
\arguments{
\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}

\item{mode}{1: KPSS-PSR+ERS, 2 - KPSS-BOOTSTRAP+ERS, 3: ERS-PSR,
4: ERS-BOOTSTRAP, 5: KPSS-PSR, 6: KPSS-BOOTSTRAP}
}
\value{
Test   (KPPS, PSR, ERS, BOOTSTRAP)
        Result (0: Non-Stationary, 1: Stationary, NA: Unable to test)
}
\description{
Decision tree 2-steps stationarity test.
}

