% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt3steps.test.R
\name{dt3steps.test}
\alias{dt3steps.test}
\title{Decision Tree 3-Steps Stationarity Test}
\usage{
dt3steps.test(data, alpha)
}
\arguments{
\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}
}
\value{
Test   (KPPS, PSR, ERS, BOOTSTRAP)
        Result (0: Non-Stationary, 1: Stationary, NA: Unable to test)
}
\description{
Decision tree 3-steps stationarity test.
}

