% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliot.rothenberg.stock.test.R
\name{elliot.rothenberg.stock.test}
\alias{elliot.rothenberg.stock.test}
\title{Elliot-Rothenberg-Stock (ERS) Stationarity Test}
\usage{
elliot.rothenberg.stock.test(data, alpha)
}
\arguments{
\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}
}
\value{
0: Non-Stationary, 1: Stationary, NA: Unable to test
}
\description{
Elliot-Rothenberg-Stock stationarity test.
}
