% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enders.test.R
\name{enders.test}
\alias{enders.test}
\title{Enders sequential test procedure for unit roots}
\usage{
enders.test(data, alpha)
}
\arguments{
\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}
}
\value{
0: Non-Stationary, 1: Stationary, 2: Stationary around non-zero mean,
3: Stationary around trend
}
\description{
Procedure to test for unit roots.
}
