% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ERROR_E}
\alias{ERROR_E}
\title{Double Exponential error}
\format{An object of class \code{numeric} of length 1.}
\usage{
ERROR_E
}
\description{
Double Exponential error
}
\keyword{datasets}
