% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.seeds.R
\name{get.seeds}
\alias{get.seeds}
\title{Seed Generator}
\usage{
get.seeds(N, filename = NULL)
}
\arguments{
\item{N}{Number of seeds to upload or generate}

\item{filename}{Name of the file to load the seeds from (Default: NULL)}
}
\value{
Vector with N seeds
}
\description{
Uploads or generates random seeds. The file must have a single seed per row
and no column header.
}
\examples{
get.seeds(1000)

}
