% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolmogorov.smirnov.test.R
\name{kolmogorov.smirnov.test}
\alias{kolmogorov.smirnov.test}
\title{Kolmogorov-Smirnov stability algorithm}
\usage{
kolmogorov.smirnov.test(data, alpha)
}
\arguments{
\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}
}
\value{
0: Non-Stationary, 1: Stationary
}
\description{
Procedure to test for time series stability described using the
Kolmorov-Smirnov test
}
