% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwiatkowski.phillips.schmidt.shin.test.R
\name{kwiatkowski.phillips.schmidt.shin.test}
\alias{kwiatkowski.phillips.schmidt.shin.test}
\title{Kwiatkowski–Phillips–Schmidt–Shin (KPSS) Stationarity Test}
\usage{
kwiatkowski.phillips.schmidt.shin.test(data, alpha)
}
\arguments{
\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}
}
\value{
0: Non-Stationary, 1: Stationary, NA: Unable to test
}
\description{
Kwiatkowski–Phillips–Schmidt–Shin stationarity test.
}

