% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ljung.box.test.R
\name{ljung.box.test}
\alias{ljung.box.test}
\title{Ljung-Box (LB) Stationarity Test}
\usage{
ljung.box.test(data, alpha)
}
\arguments{
\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}
}
\value{
0: Non-Stationary, 1: Stationary, NA: Unable to test
}
\description{
Ljung-Box stationarity test.
}

