% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phillips.perron.test.R
\name{phillips.perron.test}
\alias{phillips.perron.test}
\title{Phillips-Perron (PP) Stationarity Test}
\usage{
phillips.perron.test(data, alpha)
}
\arguments{
\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}
}
\value{
0: Non-Stationary, 1: Stationary, NA: Unable to test
}
\description{
Phillips-Perron stationarity test.
}

