% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priestley.subba.rao.test.R
\name{priestley.subba.rao.test}
\alias{priestley.subba.rao.test}
\title{Priestley-Subba Rao (PSR) Stationarity Test}
\usage{
priestley.subba.rao.test(data, alpha)
}
\arguments{
\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}
}
\value{
0: Non-Stationary, 1: Stationary, NA: Unable to test
}
\description{
Priestley-Subba Rao stationarity test.
}
