% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.test.R
\name{run.test}
\alias{run.test}
\title{Run Time Series Stationarity Test}
\usage{
run.test(test, data, alpha)
}
\arguments{
\item{test}{Name of the time series stationarity test}

\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}
}
\value{
0: Non-Stationary, 1: Stationary, 2: Stationary around non-zero mean,
3: Stationary around trend, NA: Unable to perform the test
}
\description{
Performs a time series stationarity test.
}
