% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schmidt.phillips.test.R
\name{schmidt.phillips.test}
\alias{schmidt.phillips.test}
\title{Schmidt-Phillips (SP) Stationarity Test}
\usage{
schmidt.phillips.test(data, alpha)
}
\arguments{
\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}
}
\value{
0: Non-Stationary, 1: Stationary, NA: Unable to test
}
\description{
Schmidt-Phillips stationarity test.
}
