% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sta.test.R
\name{sta.test}
\alias{sta.test}
\title{Sequential Testing Algorithm (STA)}
\usage{
sta.test(data, alpha, window)
}
\arguments{
\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}

\item{window}{Structural change window size proportion (0 - 1)}
}
\value{
ERS, BG, AT, SC, Total results (0: Non-Stationary, 1: Stationary,
NA: Not used test)
}
\description{
Sequential testing algorithm for stationarity testing.
}
