% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationarityTest.R
\name{stationarityTest}
\alias{stationarityTest}
\title{Stationarity test using the multitaper spectrum estimation method}
\usage{
stationarityTest(X, alpha = 0.05, method = "parametric")
}
\arguments{
\item{X}{Vector of 1D time series to test, whose length is at least 40}

\item{alpha}{Float of significance level (between 0 and 1, default to 0.05)}

\item{method}{String of method type ('nonparametric' or 'parametric', default to 'parametric'), 
'nonparametric' means two-way ANOVA test and 'parametric' means rank-based test}
}
\value{
Boolean (False means rejection of the null hypothesis that the time seres is stationary)
}
\description{
This function uses the multitaper method to construct hypothesis tests to test whether
a given one dimensional is stationary. The test can be either two-way ANOVA or Rank-based test
}
\examples{
data(ppp)
dat <- ppp$japan
plot(dat)
stationarityTest(X=dat, alpha=0.01, method='parametric')
plot(diff(log(dat)))
stationarityTest(X=diff(log(dat)), alpha=0.01, method='parametric')

}
