% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{STATIONARY}
\alias{STATIONARY}
\title{Stationary constant}
\format{An object of class \code{numeric} of length 1.}
\usage{
STATIONARY
}
\description{
Stationary constant
}
\keyword{datasets}

