% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{STATIONARY_TREND}
\alias{STATIONARY_TREND}
\title{Stationary around trend}
\format{An object of class \code{numeric} of length 1.}
\usage{
STATIONARY_TREND
}
\description{
Stationary around trend
}
\keyword{datasets}
