% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.change.test.R
\name{structure.change.test}
\alias{structure.change.test}
\title{Structure Change (SC) Stationarity Test}
\usage{
structure.change.test(data, alpha, type, window)
}
\arguments{
\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}

\item{type}{Type of fluctuation process "ME", "OLS-CUSUM", "OLS-MOSUM"}

\item{window}{Percent of time series size to use as window}
}
\value{
0: Non-Stationary, 1: Stationary, NA: Unable to test
}
\description{
Strcuture Change stationarity test.
}

