% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enders.test.R
\name{tau.tau.critical.function}
\alias{tau.tau.critical.function}
\title{Tau.tau critical value}
\usage{
tau.tau.critical.function(size)
}
\arguments{
\item{size}{Sample size}
}
\description{
Use the DF table for critical values to calculate the t-critical value
based on sample size. Use Table A in Enders (2009) and Dickey-Fuller (1976).
}
\note{
Includes only values for significance 0.05

Uses interpolation to calculate the t.critical

Uses maximum sample size of 1000 to calculate the linear interpolation
in the case sample size greater than 500
}

