% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.break.mean.R
\name{ts.break.mean}
\alias{ts.break.mean}
\title{Time Series Data Generator with Break in the Mean}
\usage{
ts.break.mean(N, TS, deltas, phi, theta, error, seeds, burnin)
}
\arguments{
\item{N}{Number of time series}

\item{TS}{Size of the time series}

\item{deltas}{Vector with two mean values}

\item{phi}{Vector of autoregressive parameters}

\item{theta}{Vector of moving average parameters}

\item{error}{Type of error and parameters
Normal      - c(ERROR_N, mean, stdv)
Exponential - c(ERROR_E, mean, lambda)
Triangle    - c(ERROR_T, lower, upper, mode)}

\item{seeds}{Vector of the seeds}

\item{burnin}{Number of samples thrown away at the beginning of time series generation}
}
\value{
N time series of size TS
}
\description{
Generates non-stationary time series data with break in the mean and
normal error distribution.
}
\examples{
ts.break.mean(5, 5000, c(0, 2), 0.9, 0, c(ERROR_N, 0, 1), c(645,983,653,873,432), 10)

}
