% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.break.variance.R
\name{ts.break.variance}
\alias{ts.break.variance}
\title{Time Series Data Generator with Break in the Variance}
\usage{
ts.break.variance(N, TS, delta, phi, theta, error, seeds, burnin)
}
\arguments{
\item{N}{Number of time series}

\item{TS}{Size of the time series}

\item{delta}{Mean}

\item{phi}{Autoregressive parameter}

\item{theta}{Moving average parameter}

\item{error}{Type of error and parameters
Normal      - c(ERROR_N, mean, stdv[1], stdv[2])
Exponential - c(ERROR_E, mean, lambda[1], lambda[2])
Triangle    - c(ERROR_T, lower, upper, mode[1], mode[2])}

\item{seeds}{Vector of seeds}

\item{burnin}{Number of samples thrown away at the beginning of time series generation}
}
\value{
N time series of size TS
}
\description{
Generates non-stationary time series data with break in the variance and
normal error distribution.
}
\examples{
ts.break.variance(5, 5000, 0, 0.9, 0, c(ERROR_N, 0, 1, 2), c(645,983,653,873,432), 10)

}
