% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.data.generator.R
\name{ts.data.generator}
\alias{ts.data.generator}
\title{Time Series Data Generator}
\usage{
ts.data.generator(N, delta, tau, phi, theta, mu, sigma, omega, drop = 0)
}
\arguments{
\item{N}{Number of samples}

\item{delta}{Mean}

\item{tau}{Trend on the mean}

\item{phi}{Matrix of autoregressive parameters over time}

\item{theta}{Matrix of moving average parameters over time}

\item{mu}{Normal error mean}

\item{sigma}{Normal error standard deviation}

\item{omega}{Trend on the variance}

\item{drop}{Number of initial samples to drop (Default: 0)}
}
\value{
Time series of size N
}
\description{
Generates stationary or non-stationary time series data with mean or
variance trend and normal error distribution.
}
\examples{
ts.data.generator(100, 0, 0, 1, 0, 0, 1, 0)

}

