% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.stationary.normal.R
\name{ts.stationary.normal}
\alias{ts.stationary.normal}
\title{Time Series Data Generator with Normal Error Distribution}
\usage{
ts.stationary.normal(N, TS, phi, theta, mu, sigma, seeds)
}
\arguments{
\item{N}{Number of time series}

\item{TS}{Size of the time series}

\item{phi}{Vector of autoregressive parameters}

\item{theta}{Vector of moving average parameters}

\item{mu}{Normal error mean}

\item{sigma}{Normal error standard deviation}

\item{seeds}{Vector of the seeds}
}
\value{
N time series of size TS
}
\description{
Generates non-stationary time series data with normal error distribution.
}
\examples{
ts.stationary.normal(5, 5000, 0.9, 0, 0, 1, c(645,983,653,873,432))

}

