% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.trend.autocorrelation.R
\name{ts.trend.autocorrelation}
\alias{ts.trend.autocorrelation}
\title{Time Series Data Generator with Autocorrelation Trend}
\usage{
ts.trend.autocorrelation(N, TS, delta, phis, thetas, error, seeds, burnin)
}
\arguments{
\item{N}{Number of time series}

\item{TS}{Size of the time series}

\item{delta}{Mean}

\item{phis}{Vector with the minimum and maximum autoregressive values}

\item{thetas}{Vector with the minimum and maximum moving average values}

\item{error}{Type of error and parameters
Normal      - c(ERROR_N, mean, stdv)
Exponential - c(ERROR_E, mean, lambda)
Triangle    - c(ERROR_T, lower, upper, mode)}

\item{seeds}{Vector of the seeds}

\item{burnin}{Number of samples thrown away at the beginning of time series generation}
}
\value{
N time series of size TS
}
\description{
Generates non-stationary time series data with autocorrelation trend and
normal error distribution.
}
\examples{
ts.trend.autocorrelation(5, 5000, 0, c(-0.9, 0.9), c(0, 0), c(ERROR_N, 0, 1),
c(645,983,653,873,432), 10)

}
