% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.trend.variance.R
\name{ts.trend.variance}
\alias{ts.trend.variance}
\title{Time Series Data Generator with Variance Trend}
\usage{
ts.trend.variance(N, TS, delta, phi, theta, error, omega, seeds, burnin)
}
\arguments{
\item{N}{Number of time series}

\item{TS}{Size of the time series}

\item{delta}{Mean}

\item{phi}{Vector of autoregressive parameters}

\item{theta}{Vector of moving average parameters}

\item{error}{Type of error and parameters
Normal      - c(ERROR_N, mean, stdv)
Exponential - c(ERROR_E, mean, lambda)
Triangle    - c(ERROR_T, lower, upper, mode)}

\item{omega}{Trend on the variance}

\item{seeds}{Vector of seeds}

\item{burnin}{Number of samples thrown away at the beginning of time series generation}
}
\value{
N time series of size TS
}
\description{
Generates non-stationary time series data with variance trend and
normal error distribution.
}
\examples{
ts.trend.variance(5, 5000, 0, 0.9, 0, c(ERROR_N, 0, 1), 2, c(645,983,653,873,432), 10)

}
