% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwrun.test.R
\name{wwrun.test}
\alias{wwrun.test}
\title{Grazzini stationarity test algorithm}
\usage{
wwrun.test(data, alpha, window)
}
\arguments{
\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}

\item{window}{Size of the window to perform the test}
}
\value{
0: Non-Stationary, 1: Stationary
}
\description{
Procedure to test for time series stationarity described in
Granzzini, J. (2012) Analysis of the emergent properties: Stationarity
and ergodicity, JASSS, 15:(2).
}
