% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zivot.andrews.test.R
\name{zivot.andrews.test}
\alias{zivot.andrews.test}
\title{Zivot-Andrews (ZA) Stationarity Test}
\usage{
zivot.andrews.test(data, alpha)
}
\arguments{
\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}
}
\value{
0: Non-Stationary, 1: Stationary, NA: Unable to test
}
\description{
Zivot-Andrews stationarity test.
}

