% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo.test.R
\name{algo.test}
\alias{algo.test}
\title{Sequential Test Procedure Algorithm Stationarity Test}
\usage{
algo.test(data, alpha, mode)
}
\arguments{
\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}

\item{mode}{1: ALGO-ERS, 2: ALGO-ERS+SC50, 3: ALGO-ERS+SC60,
4: ALGO-ERS+SC70, 5: ALGO-ERS+SC80, 6: ALGO-ERS+SC90}
}
\value{
0: Non-Stationary, 1: Stationary, NA: Unable to test
}
\description{
Sequential test procedure algorithm stationarity test.
}

