% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo.test.R
\name{algo.test}
\alias{algo.test}
\title{Sequential Test Procedure Algorithm Stationarity Test}
\usage{
algo.test(data, alpha, mode, window)
}
\arguments{
\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}

\item{mode}{1: ERS, 2: SP}

\item{window}{Structural change window size}
}
\value{
ERS|SP, BG, AT, SC, Total results (0: Non-Stationary, 1: Stationary,
NA: Not used test)
}
\description{
Sequential test procedure algorithm stationarity test.
}

