% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootwptos.test.R
\name{bootwptos.test}
\alias{bootwptos.test}
\title{Bootstrap Wavelet Packet Tests}
\usage{
bootwptos.test(data, alpha)
}
\arguments{
\item{data}{Time series data}

\item{alpha}{Value of alpha for the statistics test}
}
\value{
0: Non-Stationary, 1: Stationary, NA: Unable to test
}
\description{
Bootstrap wavelet packet tests.
}
