% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{NONSTATIONARY}
\alias{NONSTATIONARY}
\title{Non-Stationary constant}
\format{An object of class \code{numeric} of length 1.}
\usage{
NONSTATIONARY
}
\description{
Non-Stationary constant
}
\keyword{datasets}
