% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{STATIONARY_NONZERO_MEAN}
\alias{STATIONARY_NONZERO_MEAN}
\title{Stationary around non-zero mean}
\format{An object of class \code{numeric} of length 1.}
\usage{
STATIONARY_NONZERO_MEAN
}
\description{
Stationary around non-zero mean
}
\keyword{datasets}
<<<<<<< HEAD
=======

>>>>>>> 1980ca786f9bc64960213752d59f98e3bafdc768
