% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.break.variance.R
\name{ts.break.variance}
\alias{ts.break.variance}
\title{Time Series Data Generator with Break in the Variance}
\usage{
ts.break.variance(N, TS, delta, phi, theta, mu, sigmas, seeds, burnin)
}
\arguments{
\item{N}{Number of time series}

\item{TS}{Size of the time series}

\item{delta}{Mean}

\item{phi}{Vector of autoregressive parameters}

\item{theta}{Vector of moving average parameters}

\item{mu}{Normal error mean}

\item{sigmas}{Vector with two error standard deviation values}

\item{seeds}{Vector of the seeds}

\item{burnin}{Number of samples thrown away at the beginning of time series generation}
}
\value{
N time series of size TS
}
\description{
Generates non-stationary time series data with break in the variance and
normal error distribution.
}
\examples{
ts.break.variance(5, 5000, 0, 0.9, 0, 0, c(1, 2), c(645,983,653,873,432), 10)

}

