% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.data.generator.R
\name{ts.data.generator}
\alias{ts.data.generator}
\title{Time Series Data Generator}
\usage{
ts.data.generator(TS, y0, delta, tau, phi, theta, mu, sigma, omega, burnin)
}
\arguments{
\item{TS}{Size of the time series}

\item{y0}{Y[0] value}

\item{delta}{Drift constant}

\item{tau}{Trend on the mean}

\item{phi}{Matrix of autoregressive parameters over time}

\item{theta}{Matrix of moving average parameters over time}

\item{mu}{Normal error mean}

\item{sigma}{Normal error standard deviation}

\item{omega}{Trend on the variance}

\item{burnin}{Number of samples thrown away at the beginning of time series generation}
}
\value{
Time series of size N
}
\description{
Generates stationary or non-stationary time series data with mean or
variance trend and normal error distribution.
}
\examples{
x <- 100
ts.data.generator(x, 0, 0, 0, 1, 0, 0, 1, 0, 10)

}

