% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.data.generator.R
\name{ts.data.generator}
\alias{ts.data.generator}
\title{Time Series Data Generator}
\usage{
ts.data.generator(TS, y0, delta, tau, phi, theta, error, omega, burnin,
  continue)
}
\arguments{
\item{TS}{Size of the time series}

\item{y0}{Y[0] value}

\item{delta}{Drift constant}

\item{tau}{Trend on the mean}

\item{phi}{Matrix of autoregressive parameters over time}

\item{theta}{Matrix of moving average parameters over time}

\item{error}{Type of error and parameters
Normal      - c(ERROR_N, mean, stdv)
Exponential - c(ERROR_E, mean, lambda)
Triangle    - c(ERROR_T, lower, upper, mode)}

\item{omega}{Trend on the variance}

\item{burnin}{Number of samples thrown away at the beginning of time series generation}

\item{continue}{Continues a previous time series (Supports only AR(1) and NO MOVING AVERAGE)}
}
\value{
Time series of size N
}
\description{
Generates stationary or non-stationary time series data with mean or
variance trend and different error distributions.
}
\examples{
x <- 100
ts.data.generator(x, 0, 0, 0, 1, 0, c(ERROR_N,0,1), 0, 10, TRUE)

}
