% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.trend.autocorrelation.R
\name{ts.trend.autocorrelation}
\alias{ts.trend.autocorrelation}
\title{Time Series Data Generator with Autocorrelation Trend}
\usage{
ts.trend.autocorrelation(N, TS, delta, phis, thetas, mu, sigma, seeds)
}
\arguments{
\item{N}{Number of time series}

\item{TS}{Size of the time series}

\item{delta}{Mean}

\item{phis}{Vector with the minimum and maximum autoregressive values}

\item{thetas}{Vector with the minimum and maximum moving average values}

\item{mu}{Normal error mean}

\item{sigma}{Normal error standard deviation}

\item{seeds}{Vector of the seeds}
}
\value{
N time series of size TS
}
\description{
Generates non-stationary time series data with autocorrelation trend and
normal error distribution.
}
\examples{
ts.trend.autocorrelation(5, 5000, 0, c(-0.9, 0.9), c(0, 0), 0, 1, c(645,983,653,873,432))

}

