% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.trend.variance.R
\name{ts.trend.variance}
\alias{ts.trend.variance}
\title{Time Series Data Generator with Variance Trend}
\usage{
ts.trend.variance(N, TS, delta, phi, theta, mu, sigma, omega, seeds,
  drop = 100)
}
\arguments{
\item{N}{Number of time series}

\item{TS}{Size of the time series}

\item{delta}{Mean}

\item{phi}{Vector of autoregressive parameters}

\item{theta}{Vector of moving average parameters}

\item{mu}{Normal error mean}

\item{sigma}{Normal error standard deviation}

\item{omega}{Trend on the variance}

\item{seeds}{Vector of the seeds}

\item{drop}{Number of initial samples to drop (Default: 100)}
}
\value{
N time series of size TS
}
\description{
Generates non-stationary time series data with variance trend and
normal error distribution.
}
\examples{
ts.trend.variance(5, 5000, 0, 0.9, 0, 0, 1, 2, c(645,983,653,873,432))

}

